+++ tello.go
@@ -72,7 +72,7 @@ type Tello struct {
 	autoXY                         bool         // flag for XY autoflight
 	homeValid                      bool         // has an home point been set?
 	homeX, homeY                   float32      // set on request to provide a frame of reference
-	homeYaw                        int16        // 0 - 360 degrees, yaw when origin set
+	homeYaw                        float32      // 0 - 360 degrees, yaw when origin set
 }
 
 // ControlConnect attempts to connect to a Tello at the provided network addr.
